PretrainedTokenizer::truncateHelper: prevent array_slice() error
with the if-clause in PretrainedTokenizer::truncateHelper
certain input may result in the following error:

    array_slice(): Argument #1 ($array) must be of type array,
    null given

two tests were added to prove that the fix is working:

1. SummarizationPipelineTest:
Integration test which checks behavior using a real model
and some extracted text from a PDF. I think there is a better
way to accomplish the same test result, because this one test
runs 10+ sec. locally.

2. PretrainedTokenizerTest:
Unit test to check PretrainedTokenizer::truncateHelper itself.
The input is flawed by design, which would trigger the error
without the fix.